pdf_service/requirements.txt
flask==2.2.3
reportlab==4.0.4
gunicorn==21.2.0
python-dotenv==1.0.0
